+++ src/protocols/reading.rs
@@ -1,6 +1,6 @@
 use crate::{
     protocols::{ProtocolHandler, ReturnableConnection},
-    Connection, Node, Pea2Pea,
+    Node, Pea2Pea,
 };
 
 #[cfg(doc)]
@@ -11,7 +11,6 @@ use bytes::BytesMut;
 use futures_util::StreamExt;
 use tokio::{
     io::AsyncRead,
-    net::tcp::OwnedReadHalf,
     sync::{mpsc, oneshot},
     time::sleep,
 };
@@ -52,7 +51,7 @@ where
             while let Some((mut conn, conn_returner)) = conn_receiver.recv().await {
                 let addr = conn.addr();
                 let codec = self_clone.codec(addr);
-                let reader = self_clone.take_reader(&mut conn);
+                let reader = conn.reader.take().unwrap();
                 let framed = FramedRead::new(reader, codec);
                 let mut framed = self_clone.map_codec(framed, addr);
 
@@ -161,13 +160,6 @@ where
 
     /// Processes an inbound message. Can be used to update state, send replies etc.
     async fn process_message(&self, source: SocketAddr, message: Self::Message) -> io::Result<()>;
-
-    /// Obtains the read half of the connection to be used in [`Reading::enable_reading`].
-    fn take_reader(&self, conn: &mut Connection) -> OwnedReadHalf {
-        conn.reader
-            .take()
-            .expect("Connection's reader is not available!")
-    }
 }
 
 /// A wrapper [`Decoder`] that also counts the inbound messages.